Confusion in file naming. (#101)
Resolving the file naming and the operations to be
handled by the respective modules.

Fixes: #100
