--- test_runner.py
@@ -1,6 +1,0 @@
-"""
-This is the test runner for the Redant test automation framework. The purpose
-of the test tunner is to take a given list of servers and the TCs. Taking the
-list of TCs, create threads and pass the corresponding server and client list
-to a specific thread runner.
-"""
